Allow falsey parameters to be passed to the api
When setting a falsey parameter (Such as '{}' or '[]') to this module
the parameter gets effectively ignored due to its falsey nature. This
means that you cannot remove parameters such as 'filters' from a record
using this module (I.E. you cannot set filters to '[]' once it has been
set to anything else). I believe the intended behavior here is to not
modify parameters that are not set (Thus the != None). You you may omit
a parameter and retain the present functionality.

This may or may not be a desired change for this module overall, but it
was necessary for our use.